mulesoft-docgen/requirements.txt
lxml>=4.9.3
jinja2>=3.1.2
pathlib>=1.0.1
pyyaml>=6.0
pygments>=2.15.1
requests>=2.28.2
rich>=13.3.5 